refactor(chain): extract evm-adapter free functions into sibling modules
Phase 1 of the evm-adapter.ts structural split (zero behavior / public-API
changes). Moves module-scope free functions, constants and interfaces out
of the god-class file into cohesive siblings; the class imports back what
it uses and previously-exported symbols are re-exported from
evm-adapter.ts so the public surface is unchanged.

New files:
- evm-adapter-abi.ts        (loadAbi + require/__dirname setup)
- evm-adapter-errors.ts     (errorMessage/Code/Status, decode/enrichEvmError,
                             error/PCA interfaces, ERROR_ABI_CONTRACTS,
                             HUB_STALE_ERROR_MARKERS, isTooLowAllowanceError)
- evm-adapter-rpc.ts        (sleep, withTimeout, resolveRpcUrls,
                             boundedRetryFetchRequest, isRetryableRpcError,
                             assertSuccessfulReceipt, isKnownTransactionError)
- evm-adapter-allowance.ts  (effectivePublishAllowance, computeApprovalAction,
                             V10_PUBLISH_ONCHAIN_MIN_ALLOWANCE)
- evm-adapter-types.ts      (EVMAdapterConfig/BaseConfig, ContractCache,
                             formatProviderContext)

evm-adapter.ts: 4878 -> 4350 lines.
